+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - **Breaking**: Moved `startTime` and `endTime` for history queries to a `timeFilter` property as both or neither must be passed; passing only one parameter is not supported.
+- Renamed and promote the usage of `generateSymmetricKey()` to generate random symmetric keys.
 
 ### Fixed
 - Buffer concat error when using symmetric encryption in the browser.